Corrected Totals Reworked routes
