+++ ShepardsPies-Client/src/Components/ApplicationViews.jsx
@@ -1,10 +1,12 @@
-import { ViewAllOrders } from "./orders/ViewAllOrders";
 import { Route, Routes } from "react-router-dom";
 import { AuthorizedRoute } from "./auth/AuthorizedRoute";
+import { ViewOrders } from "./orders/ViewOrders";
 import { CreateAnOrder } from "./orders/CreateAnOrder";
 import Register from "./auth/Register";
 import Login from "./auth/Login";
 import { OrderDetails } from "./orders/OrderDetails";
+import { PizzaForm } from "./pizza/PizzaForm"
+
 
 export default function ApplicationViews({ loggedInUser, setLoggedInUser }) {
   return (
@@ -13,19 +15,23 @@ export default function ApplicationViews({ loggedInUser, setLoggedInUser }) {
 
       <Route path="/">
         <Route index path="orders" element={<AuthorizedRoute loggedInUser={loggedInUser}>
-          <ViewAllOrders />
+          <ViewOrders />
         </AuthorizedRoute>}></Route>
 
         <Route path="createorder" element={<AuthorizedRoute loggedInUser={loggedInUser}>
           <CreateAnOrder />
         </AuthorizedRoute>}></Route>
 
-        <Route path="createorder/:orderId" element={<AuthorizedRoute loggedInUser={loggedInUser}>
-          <CreateAnOrder />
+        <Route path="pizzaform/:orderId" element={<AuthorizedRoute loggedInUser={loggedInUser}>
+          <PizzaForm />
         </AuthorizedRoute>}></Route>
 
-           <Route path="orderdetails" element={<AuthorizedRoute loggedInUser={loggedInUser}>
-          <OrderDetails/>
+        <Route path="orderdetails" element={<AuthorizedRoute loggedInUser={loggedInUser}>
+          <OrderDetails />
+        </AuthorizedRoute>}></Route>
+
+         <Route path="pizzaform" element={<AuthorizedRoute loggedInUser={loggedInUser}>
+          <PizzaForm />
         </AuthorizedRoute>}></Route>
 
       </Route>
