+++ ShepardsPies-Client/src/Components/orders/CreateAnOrder.jsx
@@ -1,36 +1,114 @@
 import { useState, useEffect } from "react";
-import { useNavigate } from "react-router-dom";
-import { CreateOrder } from "../../Managers/orderManager";
-import { PizzaForm } from "../pizza/PizzaForm";
+import { GetOrders, CreateOrder } from "../../Managers/orderManager";
 import { GetEmployees } from "../../Managers/employeeManager";
-import { useParams } from "react-router-dom";
+import { GetCustomers } from "../../Managers/customerManager";
+import { useNavigate } from "react-router-dom";
 
 export const CreateAnOrder = () => {
-  const [order, setOrder] = useState({
-    customerId: null,
-    takenByEmployeeId: null,
-    deliveredByEmployeeId: null,
-    tableNumber: null,
-    tipAmount: null,
-    totalCost: null
-  });
-  const [employees, setEmployees] = useState();
-
   const navigate = useNavigate();
+  const [orders, SetOrders] = useState([]);
+  const [customers, setCustomers] = useState([]);
+  const [tableNumber, setTableNumber] = useState("");
+  const [tipAmount, setTipAmount] = useState("");
+
+  const [employees, setEmployees] = useState([]);
+  const [takenByEmployee, setTakenByEmployee] = useState("");
+  const [deliveryByEmployee, setDeliveryByEmployee] = useState("");
+  const [selectedCustomerId, setSelectedCustomerId] = useState("");
+  const toNullableInt = (val) => (val === "" ? null : parseInt(val));
+  const toNullableDecimal = (val) => val === "" ? null : parseFloat(val);
+
+  const getAllOrders = () => {
+    GetOrders().then(SetOrders);
+  };
 
   useEffect(() => {
-    GetEmployees().then(setEmployees)
+    getAllOrders();
   }, []);
 
+  useEffect(() => {
+    GetEmployees().then(setEmployees);
+  }, []);
 
+  useEffect(() => {
+    GetCustomers().then(setCustomers);
+  }, []);
 
+  const handleOrderCreate = () => {
+    const order = {
+      TableNumber: toNullableInt(tableNumber),
+      CustomerId: toNullableInt(selectedCustomerId),
+      TakenByEmployeeId: toNullableInt(takenByEmployee),
+      DeliveredByEmployeeId: toNullableInt(deliveryByEmployee),
+      TipAmount: toNullableDecimal(tipAmount),
+      TotalCost: toNullableDecimal(tipAmount)
+    };
+
+    CreateOrder(order).then(() => navigate("/orderdetails"));
+  };
 
   return (
     <>
       <div>
-      </div>
-      <PizzaForm></PizzaForm>
-      <button>Cancel Order</button>
+        <h3>Create New Order</h3>
+        <label>Taken by employee:</label>
+        <select
+          value={takenByEmployee}
+          onChange={(e) => setTakenByEmployee(e.target.value)}
+        >
+          <option value="">-- Select Employee --</option>
+          {employees.map((employee) => (
+            <option key={employee.id} value={employee.id}>
+              {employee.name}
+            </option>
+          ))}
+        </select>
+
+        <label>Delivery employee:</label>
+        <select
+          value={deliveryByEmployee}
+          onChange={(e) => setDeliveryByEmployee(e.target.value)}
+        >
+          <option value="">-- Select Employee --</option>
+          {employees.map((employee) => (
+            <option key={employee.id} value={employee.id}>
+              {employee.name}
+            </option>
+          ))}
+        </select>
+
+        <label>Customer Name:</label>
+        <select
+          value={selectedCustomerId}
+          onChange={(e) => setSelectedCustomerId(e.target.value)}
+        >
+          <option value="">-- Select Customer --</option>
+          {customers.map((customer) => (
+            <option key={customer.id} value={customer.id}>
+              {customer.name}
+            </option>
+          ))}
+        </select>
+
+        <label>Table Number:</label>
+        <input
+          type="number"
+          value={tableNumber}
+          onChange={(e) => setTableNumber(e.target.value)}
+          placeholder="Enter table number"
+        />
+
+        <label>Tip Amount:</label>
+        <input
+          type="number"
+          step="any"
+          value={tipAmount}
+          onChange={(e) => setTipAmount(e.target.value)}
+          placeholder="Enter tip (e.g. 4.25)"
+        />
+
+        <button onClick={handleOrderCreate}>Create Order</button>
+     </div>
     </>
   );
 };
