+++ ShepardsPies-Client/src/Managers/orderManager.js
@@ -15,3 +15,12 @@ return fetch(apiUrl, {
 }).then((res) => res.json());
 };
 
+export const updateOrderCost = (orderId) => {
+  return fetch(`${apiUrl}/${orderId}/recalculate-total`, {
+    method: "PUT",
+    headers: {
+      "Content-Type": "application/json",
+    },
+    body: JSON.stringify(orderId),
+  });
+};
